feat(modularization): S-13b -- extract federation page to web/modules/federation.js
Moves the entire Federation page (loadFederationPage, renderFederationPage,
wireFederationPage, fedOpenPeerModal, fedSavePeerModal, fedRevealToken,
fedRotateToken, fedToggleShareCap, fedDeletePeer, fedApplyToMainAgent,
fedRefreshAndReload) from app.js into web/modules/federation.js.

Module imports: t, showToast, getFederatedPeerStatus, setFederatedPeerStatus,
getChatSelectedAgent, setChatSelectedAgent. DI via initFederation({ openModal,
closeModal }).

Updated federation-ui-contract.test.ts to read FEDERATION_MOD instead of APP
for all federation-specific assertions.

Smoke: 0 pageerror, federation page renders (federationStats/Master/Peers DOM
elements visible). vitest: 2873 tests green.
